I changed the call to GetCellInfo() and simplified it removing the const double argument, since there was no point in still having it if I can oompare the cell center with PI and alter it accordingly.
I've also changed the DICEOLD #if/#else in order to use normal flow control, i.e., ansi C if/else. This is better since variable initialization at the begginning of the file was giving some warnings because such variable(s) is only used in one of the rules (DICE old or new).

Some comments were added in order to increase documentation of this program and some #includes have been transposed from the header file (portable.h) to here, to simplify the header and because their natural place is here. The other headers should not care, for example, if I use <math.h> here. This way the file got a bit bigger.

The defines for eta values were changed to const double's as mentioned at the log of portable.h


git-svn-id: file:///Users/andre/andre_msc/trunk/calo-preproc@28 b2e9302d-b4d2-0310-80c4-8586bce995ec